requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==24.3.0
auth0-python==4.7.2
Authlib==1.3.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.13.11
coloredlogs==15.0.1
crewai==0.95.0
crewai-tools==0.32.0
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.126
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.7
fastavro==1.10.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.21
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
lancedb==0.18.0
langchain==0.3.15
langchain-cohere==0.3.5
langchain-community==0.3.15
langchain-core==0.3.31
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.1.147
linkup-sdk==0.2.2
litellm==1.59.5
load-dotenv==0.1.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.48
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.2
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patronus==0.0.17
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==2.10.1
posthog==3.9.3
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.22.0
pyOpenSSL==24.3.0
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.392.post0
pysbd==0.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.13.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
scrapegraph_py==1.10.2
selenium==4.28.1
serpapi==0.1.5
setuptools==75.8.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.1.26
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.2
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.5.23
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
weaviate-client==4.10.4
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
pinecone==6.0.2
